requirements.txt
openai
yfinance
pandas
psycopg2
python-dotenv
numpy
ta
pytz
streamlit
transformers
torch
newsapi-python
fredapi
groq
psutil
plotly
tiktoken
sqlalchemy
schedule
matplotlib